moves v3 repositories to root level
