+++ app/Services/PostService.php
@@ -7,7 +7,7 @@ use Rogue\Jobs\SendPostToCustomerIo;
 use Rogue\Jobs\SendPostToQuasar;
 use Rogue\Jobs\SendDeletedPostToQuasar;
 use Rogue\Jobs\SendReviewedPostToCustomerIo;
-use Rogue\Repositories\Three\PostRepository;
+use Rogue\Repositories\PostRepository;
 
 class PostService
 {
